requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
google-generativeai==0.3.1
langchain==0.1.0
langchain-google-genai==0.0.5
chromadb==0.4.18
faiss-cpu==1.7.4
numpy==1.26.2
pandas==2.1.3
python-dotenv==1.0.0
aiohttp==3.9.1
scikit-learn==1.3.2
